Eigen::Matrix4f -> Affine3d
This conforms to most of the *_conversions functions
and slightly simplifies some things.

+++ include/curious_table_explorer/model.h
@@ -30,7 +30,7 @@ public:
 
 	void addView(const ModelView&);
 
-	Eigen::Vector4f center() const;
+	Eigen::Vector4d center() const;
 
 	const PointCloud::Ptr& convexHullPoints() const;
 	const std::vector<pcl::Vertices>& convexHullVertices() const;
